Support adding feeds from local files
Given a file:///feed.rss, Feeds is unable to open it and add the feed.
It will in some cases be possible to extract the correct url by looking
for the <atom:link rel="self" href="$url"/> that may occur in the file.

This patch will try to extract the url, and allow the user to add
the feed if the url is found, otherwise ignore it (which is the
current behaviour).